Create with-clause-1
The statement above was created to solve a specific issue: the data measured by the energy meters was showing cumulative values, while our system needed the variation that was measured each minute so it could remove possible outliers.

With that in mind 2 CTEs where generated in which one had a meter reading superior to the other. A 'period' CTE was created to retrieve all the possible dates in the collected data, while a CTE to generate all possible dates in a given time interval 
of 1 minute.

The last 'select' does the subtraction operations to build to report that, in this case, is a 'Gauge' to analyse the data consumption of a generator.